refactor(cli): extract shared obs::install from run (reused by schedule)
